cards controller created

+++ src/main/java/com/teamwebsoft/repositories/CardsRepository.java
@@ -8,5 +8,5 @@ import java.util.List;
 
 @Repository
 public interface CardsRepository extends JpaRepository<Cards, Long> {
-    List<Cards> findByCustomerId(int customerId);
+    List<Cards> findByCustomerId(long customerId);
 }
